+++ tools/check_env.php
@@ -5,6 +5,9 @@
  * Usage: php check_env.php
  */
 
+require_once __DIR__ . '/../src/Util/CliGuard.php';
+\App\Util\CliGuard::assertCli();
+
 require __DIR__ . '/../vendor/autoload.php';
 
 use App\Config\Env;